refactor(common): INT-1500 Move `AmountTransformer` to common folder
