+++ src/common/utility/index.ts
@@ -1,3 +1,4 @@
+export { default as AmountTransformer } from './amount-transformer';
 export { default as bindDecorator } from './bind-decorator';
 export { default as createFreezeProxy, createFreezeProxies } from './create-freeze-proxy';
 export { default as CacheKeyResolver } from './cache-key-resolver';
