+++ src/order/map-to-internal-order.ts
@@ -1,8 +1,9 @@
 import { filter, find, keyBy, reduce } from 'lodash';
 
-import { AmountTransformer, LineItem } from '../cart';
+import { LineItem } from '../cart';
 import { mapToInternalLineItems } from '../cart';
 import { Checkout } from '../checkout';
+import { AmountTransformer } from '../common/utility';
 import { mapToInternalCoupon } from '../coupon';
 import { HOSTED } from '../payment';
