+++ src/payment/strategies/affirm/affirm-payment-strategy.ts
@@ -1,6 +1,6 @@
-import { AmountTransformer } from '../../../cart';
 import { CheckoutStore, InternalCheckoutSelectors } from '../../../checkout';
 import { MissingDataError, MissingDataErrorType, NotInitializedError, NotInitializedErrorType } from '../../../common/error/errors';
+import { AmountTransformer } from '../../../common/utility';
 import { Order, OrderActionCreator, OrderRequestBody } from '../../../order';
 import { OrderFinalizationNotRequiredError } from '../../../order/errors';
 import { Consignment } from '../../../shipping';
